Import keras_triplet_descriptor directly
Remove adding of the dir_ktd argument to path and rename modules under
keras_triplet_descriptor to be under the package name.
Added a line to init.sh to introduce an __init__.py file to the
keras_triplet_descriptor repository automatically after downloading so
that it is recognised as a package.