Add Agenda Support
A major update in which agenda support is added, to allow you to simply
specify your files externally in an agenda file (JSON format) by passing
the file path as an argument -f to the program.
In addition, the denoise and descriptor data loading processes were
streamlined along with this update, to avoid unnecessary loading of
generators until they are needed, and to avoid reloading generators in
betwen runs of the same agenda.