Add agendas, optimizers and remove baseline_opt
Added agendas for the unet and the baseline models which can be run as
seen.
Added dyamic optimizer code generation for the sgd and adam optimizers
and removed the baseline_opt model, which is now accessible via the
`base_opt' optimizer code.
Removed a file called custommods.py that should not have been in the
repository.