benchmark.py: Load results and auto delete
Fixed the module names for hpatches_benchmark from being under the old
keras_triplet_descriptor to being under dl2019.evaluate, as they should
be.
Added automatic saving and loading of results and a command line
parameter to allow you to skip deletion of the results and eval folders
after the evaluation is complete. The default is to delete however,
which will save space.