desc_suffix: Add denoiser prepended to suffix
The denoiser used to denoise the descriptor data is now prepended to the
desc_suffix. If clean data was used, then no additional text is
prepended to the suffix if it exists. This is also used by benchmark.py
in a new way. It is thought that if a descriptor was trained with a
particular denoiser then it should be evaluated with the same
denoiser.