Distinguish denoisertrain and denoisereval
The desc_suffix parameter has been renamed to denoisertrain and will be
used to specify the denoiser that the user wishes to use to denoise the
descriptor data prior to training.
The denoisereval parameter is evaluated from the denoiser parameters
passed e.g. model_denoise and optimizer_denoise and will be used for
both training and evaluation. If the denoisereval does not match the
denoisertrain then no further training of the descriptor will occur,
but evaluation may.