models.json: Add a logical structure and new entries
The structure of the file has been carefully updated to reflect the
workflow.
New entries have been added now that we can specify separate denoise and
evaluation models to evaluate a noisy-trained descriptor with U-Net and
DnCNN.
The baselinemse.json file has been removed as it is incorporated into
models.json now.